pyproject.toml
[tool.poetry]
name = "balance-hub"
version = "0.1.0"
description = "Async REST API for accounts and payments"
authors = ["Andrey"]
license = "MIT"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.42"}
asyncpg = "^0.30.0"
alembic = "^1.16.4"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-multipart = "^0.0.20"
slowapi = "^0.1.9"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
flake8-quotes = "^3.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 99
multi_line_output = 3
include_trailing_comma = false
force_single_line = false
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 99
extend-ignore = ["E203"]
inline-quotes = "single"
multiline-quotes = "single"
docstring-quotes = "double"
